requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pydantic
websockets
aiofiles
thefuzz
sentence-transformers
pyngrok
playwright